windows-terminal: exclude empty profile fields from matchProfiles
Windows Terminal's MatchProfilesEntry skips empty profile field values, so
"source": ".*" no longer pulls in local profiles without a source and
"commandline": ".*" no longer pulls in profiles without a command line.

Moved the pure newTabMenu resolution helpers into src/new-tab-menu.ts so they
can be tested without open-profile.tsx reading settings.json at import time,
and added regression tests for the matching and ordering rules. Tests run on
the Node built-in test runner via `npm test`, so no new dependency is needed.
